+++ clients/admin-ui/src/features/subject-request/RequestFilters.tsx
@@ -5,22 +5,21 @@ import {
   InputGroup,
   InputLeftAddon,
   InputLeftElement,
-  Select,
   Stack,
   Text,
   useToast,
 } from "@fidesui/react";
-import React from "react";
+import React, { useCallback, useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { selectToken } from "../auth";
+import DropdownCheckbox from "../common/DropdownCheckbox/DropdownCheckbox";
 import {
   CloseSolidIcon,
   DownloadSolidIcon,
   SearchLineIcon,
 } from "../common/Icon";
 import PIIToggle from "../common/PIIToggle";
-import { statusPropMap } from "../common/RequestStatusBadge";
 import {
   clearAllFilters,
   requestCSVDownload,
@@ -29,8 +28,9 @@ import {
   setRequestId,
   setRequestStatus,
   setRequestTo,
-} from "./privacy-requests.slice";
-import { PrivacyRequestStatus } from "./types";
+} from "../privacy-requests/privacy-requests.slice";
+import { PrivacyRequestStatus } from "../privacy-requests/types";
+import { SubjectRequestStatusMap } from "./constants";
 
 const useRequestFilters = () => {
   const filters = useSelector(selectPrivacyRequestFilters);
@@ -39,8 +39,20 @@ const useRequestFilters = () => {
   const toast = useToast();
   const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) =>
     dispatch(setRequestId(event.target.value));
-  const handleStatusChange = (event: React.ChangeEvent<HTMLSelectElement>) =>
-    dispatch(setRequestStatus(event.target.value as PrivacyRequestStatus));
+  const handleStatusChange = useCallback(
+    (values: string[]) => {
+      const list: PrivacyRequestStatus[] = [];
+      values.forEach((v) => {
+        SubjectRequestStatusMap.forEach((value, key) => {
+          if (key === v) {
+            list.push(value as PrivacyRequestStatus);
+          }
+        });
+      });
+      dispatch(setRequestStatus(list));
+    },
+    [dispatch]
+  );
   const handleFromChange = (event: React.ChangeEvent<HTMLInputElement>) =>
     dispatch(setRequestFrom(event?.target.value));
   const handleToChange = (event: React.ChangeEvent<HTMLInputElement>) =>
@@ -77,13 +89,8 @@ const useRequestFilters = () => {
   };
 };
 
-const StatusOption: React.FC<{ status: PrivacyRequestStatus }> = ({
-  status,
-}) => <option value={status}>{statusPropMap[status].label}</option>;
-
 const RequestFilters: React.FC = () => {
   const {
-    status,
     handleSearchChange,
     handleStatusChange,
     handleFromChange,
@@ -92,27 +99,40 @@ const RequestFilters: React.FC = () => {
     handleDownloadClick,
     id,
     from,
+    status,
     to,
   } = useRequestFilters();
+
+  const loadStatusList = (values: string[]): Map<string, boolean> => {
+    const list = new Map<string, boolean>();
+    SubjectRequestStatusMap.forEach((value, key) => {
+      let result = false;
+      if (values.includes(value)) {
+        result = true;
+      }
+      list.set(key, result);
+    });
+    return list;
+  };
+
+  // Load the status list
+  const statusList = useMemo(() => loadStatusList(status || []), [status]);
+
+  // Filter the selected status list
+  const selectedStatusList = new Map(
+    [...statusList].filter(([, v]) => v === true)
+  );
+
   return (
     <Stack direction="row" spacing={4} mb={6}>
-      <Select
-        placeholder="Select status"
-        size="sm"
+      <DropdownCheckbox
+        list={statusList}
+        selectedList={selectedStatusList}
         minWidth="144px"
-        value={status || ""}
         onChange={handleStatusChange}
-        borderRadius="md"
-      >
-        <StatusOption status="approved" />
-        <StatusOption status="complete" />
-        <StatusOption status="denied" />
-        <StatusOption status="canceled" />
-        <StatusOption status="error" />
-        <StatusOption status="in_processing" />
-        <StatusOption status="paused" />
-        <StatusOption status="pending" />
-      </Select>
+        title="Select Status"
+        tooltipPlacement="top"
+      />
       <InputGroup size="sm">
         <InputLeftElement pointerEvents="none">
           <SearchLineIcon color="gray.300" w="17px" h="17px" />